Bug fixing base code

+++ Gubbi7 Menu/ActionDeleteConfiguration.cs
@@ -20,7 +20,8 @@ namespace Gubbi7_Menu
 
 		public void Invoke()
 		{
-			configController.DeleteConfiguration(name);
+			if(MenuUtility.ConfirmAction($"slette {name}"))
+				configController.DeleteConfiguration(name);
 		}
 	}
 }
